Fix: feedback controller modification
